fix(_sedefaults): universal functions for colly
